fix: Refactor Helpers module
The Realtime.Helpers module was becoming a catch all for multiple contexts. This PR separates it into multiple contexts and moves some functions to other modules more fitting to their context.

Also we have simpler ways to connect to the database to reduce error prone operations (e.g. having to define argument one by one in a specific order)
