+++ lib/extensions/postgres_cdc_rls/replication_poller.ex
@@ -9,14 +9,18 @@ defmodule Extensions.PostgresCdcRls.ReplicationPoller do
   require Logger
 
   import Realtime.Helpers,
-    only: [cancel_timer: 1, default_ssl_param: 1, connect_db: 9, log_error: 2, to_log: 1]
+    only: [cancel_timer: 1, log_error: 2, to_log: 1]
 
   alias DBConnection.Backoff
-  alias Extensions.PostgresCdcRls.{Replications, MessageDispatcher}
-  alias Realtime.Adapters.Changes.{DeletedRecord, NewRecord, UpdatedRecord}
-  alias Realtime.PubSub
 
-  @queue_target 5_000
+  alias Extensions.PostgresCdcRls.MessageDispatcher
+  alias Extensions.PostgresCdcRls.Replications
+
+  alias Realtime.Adapters.Changes.DeletedRecord
+  alias Realtime.Adapters.Changes.NewRecord
+  alias Realtime.Adapters.Changes.UpdatedRecord
+  alias Realtime.Database
+  alias Realtime.PubSub
 
   def start_link(opts), do: GenServer.start_link(__MODULE__, opts)
 
@@ -24,20 +28,14 @@ defmodule Extensions.PostgresCdcRls.ReplicationPoller do
   def init(args) do
     tenant = args["id"]
     Logger.metadata(external_id: tenant, project: tenant)
-    ssl_enforced = default_ssl_param(args)
 
-    {:ok, conn} =
-      connect_db(
-        args["db_host"],
-        args["db_port"],
-        args["db_name"],
-        args["db_user"],
-        args["db_password"],
-        1,
-        @queue_target,
-        ssl_enforced,
-        "realtime_rls"
-      )
+    # higher number of pool connections leads to issues
+    realtime_rls_settings =
+      args
+      |> Database.from_settings("realtime_rls")
+      |> Map.put(:pool, 1)
+
+    {:ok, conn} = Database.connect_db(realtime_rls_settings)
 
     state = %{
       backoff: Backoff.new(backoff_min: 100, backoff_max: 5_000, backoff_type: :rand_exp),
