+++ lib/realtime/tenants/migrations.ex
@@ -5,6 +5,8 @@ defmodule Realtime.Tenants.Migrations do
 
   require Logger
   import Realtime.Helpers, only: [log_error: 2]
+  alias Realtime.Crypto
+  alias Realtime.Database
   alias Realtime.Repo
 
   alias Realtime.Tenants.Migrations.{
@@ -51,8 +53,6 @@ defmodule Realtime.Tenants.Migrations do
     RemoveCheckColumns,
     RedefineAuthorizationTables
   }
-
-  alias Realtime.Helpers
 
   @migrations [
     {20_211_116_024_918, CreateRealtimeSubscriptionTable},
@@ -111,16 +111,10 @@ defmodule Realtime.Tenants.Migrations do
         } = settings
       ) do
     {host, port, name, user, pass} =
-      Helpers.decrypt_creds(
-        db_host,
-        db_port,
-        db_name,
-        db_user,
-        db_password
-      )
+      Crypto.decrypt_creds(db_host, db_port, db_name, db_user, db_password)
 
-    {:ok, addrtype} = Helpers.detect_ip_version(host)
-    ssl_enforced = Helpers.default_ssl_param(settings)
+    {:ok, addrtype} = Database.detect_ip_version(host)
+    ssl_enforced = Database.default_ssl_param(settings)
 
     [
       hostname: host,
@@ -133,7 +127,7 @@ defmodule Realtime.Tenants.Migrations do
       parameters: [application_name: "realtime_migrations"],
       backoff_type: :stop
     ]
-    |> Helpers.maybe_enforce_ssl_config(ssl_enforced)
+    |> Database.maybe_enforce_ssl_config(ssl_enforced)
     |> Repo.with_dynamic_repo(fn repo ->
       Logger.info("Applying migrations to #{host}")
