+++ lib/realtime_web/live/page_live/index.html.heex
@@ -6,7 +6,7 @@
   <h3 class="font-bold">Tools</h3>
   <ul>
     <li><.link href={Routes.inspector_index_path(@socket, :new)}>🔎 Inspector</.link></li>
-    <li><.link href={Routes.inspector_index_path(@socket, :index)}>🟢 Status</.link></li>
+    <li><.link href={Routes.status_index_path(@socket, :index)}>🟢 Status</.link></li>
   </ul>
 </div>
