+++ lib/realtime_web/plugs/auth_tenant.ex
@@ -5,19 +5,18 @@ defmodule RealtimeWeb.AuthTenant do
   require Logger
 
   import Plug.Conn
-  import Realtime.Helpers
 
   alias Realtime.Api.Tenant
+  alias Realtime.Crypto
+
   alias RealtimeWeb.ChannelsAuthorization
 
   def init(opts), do: opts
 
   def call(%{assigns: %{tenant: tenant}} = conn, _opts) do
-    secure_key = Application.get_env(:realtime, :db_enc_key)
-
     with %Tenant{jwt_secret: jwt_secret, jwt_jwks: jwt_jwks} <- tenant,
          token when is_binary(token) <- access_token(conn),
-         jwt_secret_dec <- decrypt!(jwt_secret, secure_key),
+         jwt_secret_dec <- Crypto.decrypt!(jwt_secret),
          {:ok, claims} <- ChannelsAuthorization.authorize_conn(token, jwt_secret_dec, jwt_jwks) do
       Logger.metadata(external_id: tenant.external_id, project: tenant.external_id)
