+++ test/realtime_web/controllers/tenant_controller_test.exs
@@ -3,9 +3,10 @@ defmodule RealtimeWeb.TenantControllerTest do
   use RealtimeWeb.ConnCase, async: false
 
   import Mock
-  import Realtime.Helpers, only: [encrypt!: 2, transaction: 2]
 
   alias Realtime.Api.Tenant
+  alias Realtime.Crypto
+  alias Realtime.Database
   alias Realtime.PromEx.Plugins.Tenants
   alias Realtime.Tenants
   alias Realtime.Tenants.Cache
@@ -83,11 +84,10 @@ defmodule RealtimeWeb.TenantControllerTest do
         ext_id = @default_tenant_attrs["external_id"]
         conn = put(conn, Routes.tenant_path(conn, :update, ext_id), tenant: @default_tenant_attrs)
         [%{"settings" => settings}] = json_response(conn, 201)["data"]["extensions"]
-        sec_key = Application.get_env(:realtime, :db_enc_key)
-        assert encrypt!("127.0.0.1", sec_key) == settings["db_host"]
-        assert encrypt!("postgres", sec_key) == settings["db_name"]
-        assert encrypt!("supabase_admin", sec_key) == settings["db_user"]
-        assert encrypt!("postgres", sec_key) == settings["db_password"]
+        assert Crypto.encrypt!("127.0.0.1") == settings["db_host"]
+        assert Crypto.encrypt!("postgres") == settings["db_name"]
+        assert Crypto.encrypt!("supabase_admin") == settings["db_user"]
+        assert Crypto.encrypt!("postgres") == settings["db_password"]
       end
     end
 
@@ -266,7 +266,7 @@ defmodule RealtimeWeb.TenantControllerTest do
       with_mock JwtVerification, verify: fn _token, _secret, _jwks -> {:ok, %{}} end do
         {:ok, db_conn} = Connect.lookup_or_start_connection(ext_id)
 
-        transaction(db_conn, fn transaction_conn ->
+        Database.transaction(db_conn, fn transaction_conn ->
           Postgrex.query!(transaction_conn, "DROP SCHEMA realtime CASCADE", [])
           Postgrex.query!(transaction_conn, "CREATE SCHEMA realtime", [])
           Postgrex.query!(transaction_conn, "DROP ROLE supabase_realtime_admin", [])
